# data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACG Theogony API",
    "description": "ACG × 神话原型 关系知识图谱 API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/relationships": {
      "get": {
        "tags": [
          "relationships"
        ],
        "summary": "List Relationships",
        "operationId": "list_relationships_api_relationships_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "origin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Confidence"
            }
          },
          {
            "name": "characterId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Characterid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "relationships"
        ],
        "summary": "Submit Relationship",
        "description": "众包提交关系（进入审核队列）。",
        "operationId": "submit_relationship_api_relationships_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationshipSubmit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/pending": {
      "get": {
        "tags": [
          "relationships"
        ],
        "summary": "Review Pending",
        "operationId": "review_pending_api_review_pending_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "x-review-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Review-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/{relationship_id}": {
      "post": {
        "tags": [
          "relationships"
        ],
        "summary": "Review Decide",
        "operationId": "review_decide_api_review__relationship_id__post",
        "parameters": [
          {
            "name": "relationship_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Relationship Id"
            }
          },
          {
            "name": "x-review-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Review-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/batch": {
      "post": {
        "tags": [
          "relationships"
        ],
        "summary": "Review Batch",
        "description": "批量审核：{\"<id>\": \"approve\"|\"reject\", ...}",
        "operationId": "review_batch_api_review_batch_post",
        "parameters": [
          {
            "name": "x-review-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Review-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "title": "Decisions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RelationshipSubmit": {
        "properties": {
          "sourceId": {
            "type": "string",
            "title": "Sourceid"
          },
          "targetId": {
            "type": "string",
            "title": "Targetid"
          },
          "type": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "evidence": {
            "type": "string",
            "maxLength": 200,
            "title": "Evidence"
          },
          "contributor": {
            "type": "string",
            "maxLength": 32,
            "title": "Contributor",
            "default": "匿名"
          }
        },
        "type": "object",
        "required": [
          "sourceId",
          "targetId",
          "type",
          "evidence"
        ],
        "title": "RelationshipSubmit",
        "description": "众包关系提交（进入审核队列）。"
      },
      "RelationshipType": {
        "type": "string",
        "enum": [
          "PARENT_OF",
          "SIBLING_OF",
          "SPOUSE_OF",
          "LOVER_OF",
          "MASTER_OF",
          "ALLY_OF",
          "ENEMY_OF",
          "FOUGHT_WITH",
          "MENTOR_OF",
          "PROTOTYPE_IS",
          "GENDER_SWAP",
          "DERIVED_FROM",
          "COMPOSITE_OF",
          "BELONGS_TO"
        ],
        "title": "RelationshipType",
        "description": "关系类型（存储层只用规范方向，逆关系在入库时归一化）。"
      },
      "ReviewDecision": {
        "properties": {
          "action": {
            "type": "string",
            "pattern": "^(approve|reject)$",
            "title": "Action"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewDecision"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}